refactor: split openapi pipeline
